board class in separate files
